+++ fabtools/require/deb.py
@@ -11,13 +11,13 @@ from __future__ import with_statement
 from fabric.utils import puts
 
 from fabtools.deb import (
-    distrib_codename,
     install,
     is_installed,
     uninstall,
     update_index,
 )
 from fabtools.files import is_file, watch
+from fabtools.system import distrib_codename
 from fabtools.utils import run_as_root