+++ fabtools/rpm.py
@@ -10,7 +10,6 @@ from __future__ import with_statement
 
 from fabric.api import hide, run, settings
 
-from fabtools.files import is_file
 from fabtools.utils import run_as_root
 
 
@@ -170,66 +169,6 @@ def groupuninstall(group, options=None):
         options = [options]
     options = " ".join(options)
     run_as_root('%(manager)s %(options)s groupremove "%(group)s"' % locals())
-
-
-def distrib_id():
-    """
-    Get the ID of the distrib.
-
-    Example::
-
-        from fabtools.rpm import distrib_id
-
-        if distrib_id() == 'CentOS':
-            print('%s is not running RHEL.' % (env.host))
-
-    """
-    with settings(hide('running', 'stdout')):
-        return run('lsb_release --id --short')
-
-
-def distrib_codename():
-    """
-    Get the codename of the distrib.
-
-    Example::
-
-        from fabtools.rpm import distrib_codename
-
-        if distrib_codename() == 'Final':
-            print('%s is running final version of %s %s.'
-              % (env.host, distrib_id(), distrib_release))
-
-    """
-    with settings(hide('running', 'stdout')):
-        return run('lsb_release --codename --short')
-
-
-def distrib_desc():
-    """
-    Get the description of the distrib.
-
-    """
-    with settings(hide('running', 'stdout')):
-        if not is_file('/etc/redhat-release'):
-            return run('lsb_release --desc --short')
-        return run('cat /etc/redhat-release')
-
-
-def distrib_release():
-    """
-    Get the release number of the distrib.
-
-    Example::
-
-        from fabtools.rpm import distrib_release
-
-        if distrib_release() == '6.1' and distrib_id == 'CentOS':
-            print('CentOS 6.2 has been released. Please update.')
-
-    """
-    with settings(hide('running', 'stdout')):
-        return run('lsb_release -r --short')
 
 
 def repolist(status='', media=None):