+++ docs/CHANGELOG.rst
@@ -8,6 +8,7 @@ Version 0.14.0dev (unreleased)
 * Add support for Vagrant 1.1 providers in functional tests
 * Add ``fabtools.system.cpus`` to get the host's CPU count
 * Less verbose output
+* Move OS detection functions to ``fabtools.system``
 
 Version 0.13.0 (2013-03-15)
 ---------------------------